+++ scripts/build_against_sdk.sh
@@ -95,14 +95,13 @@ platform=$(echo $UNAME | sed "y/ABCDEFGHIJKLMNOPQRSTUVWXYZ/abcdefghijklmnopqrstu
 # mapnik 3.x / c++11 enabled
 if [[ ${platform} == 'linux' ]]; then
     upgrade_clang
-    HASH="2091-g68fba05"
+    HASH="2235-gb907634"
     TARBALL_NAME="mapnik-${platform}-sdk-v2.2.0-${HASH}"
 fi
 
 if [[ $platform == 'darwin' ]]; then
     platform="macosx"
-    HASH="2091-g68fba05"
-    TARBALL_NAME="mapnik-${platform}-sdk-v2.2.0-${HASH}-lto"
+    TARBALL_NAME="mapnik-${platform}-sdk-v3.0.0-rc1-5-gc3dbd83-lto"
 fi
 
 
@@ -111,10 +110,11 @@ export MAPNIK_SDK=${BUILD_DIR}/${TARBALL_NAME}
 export PATH=${MAPNIK_SDK}/bin:${PATH}
 export PKG_CONFIG_PATH=${MAPNIK_SDK}/lib/pkgconfig
 
-echo "looking for ~/projects/mapnik-packaging/osx/out/dist/${TARBALL_NAME}.${COMPRESSION}"
-if [ -f "$HOME/projects/mapnik-packaging/osx/out/dist/${TARBALL_NAME}.${COMPRESSION}" ]; then
+LOCAL_PACKAGE="$HOME/projects/mapnik-package-lto/osx/out/dist"
+echo "looking for ${LOCAL_PACKAGE}/${TARBALL_NAME}.${COMPRESSION}"
+if [ -f "${LOCAL_PACKAGE}/${TARBALL_NAME}.${COMPRESSION}" ]; then
     echo "copying over ${TARBALL_NAME}.${COMPRESSION}"
-    cp "$HOME/projects/mapnik-packaging/osx/out/dist/${TARBALL_NAME}.${COMPRESSION}" .
+    cp "${LOCAL_PACKAGE}/${TARBALL_NAME}.${COMPRESSION}" .
 else
     if [ ! -f "${TARBALL_NAME}.${COMPRESSION}" ]; then
         echo "downloading ${REMOTE_URI}"
@@ -153,6 +153,9 @@ MODULE_PATH=$(node-pre-gyp reveal module_path ${ARGS})
 rm -rf ${MODULE_PATH}
 npm install --build-from-source ${ARGS} --clang=1
 npm ls
+# copy shapeindex and nik2img
+cp ${MAPNIK_SDK}/bin/shapeindex ${MODULE_PATH}
+cp ${MAPNIK_SDK}/bin/nik2img ${MODULE_PATH}
 # copy lib
 cp ${MAPNIK_SDK}/lib/libmapnik.* ${MODULE_PATH}
 # copy plugins