+++ src/mapnik_cairo_surface.hpp
@@ -1,7 +1,11 @@
 #ifndef __NODE_MAPNIK_CAIRO_H__
 #define __NODE_MAPNIK_CAIRO_H__
 
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wunused-parameter"
 #include <nan.h>
+#pragma GCC diagnostic pop
+
 #include <string>
 #include <sstream>