+++ src/mapnik_feature.hpp
@@ -1,7 +1,11 @@
 #ifndef __NODE_MAPNIK_FEATURE_H__
 #define __NODE_MAPNIK_FEATURE_H__
 
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wunused-parameter"
 #include <nan.h>
+#pragma GCC diagnostic pop
+
 #include "mapnik3x_compatibility.hpp"
 
 // mapnik
@@ -19,12 +23,11 @@ public:
     static void Initialize(Handle<Object> target);
     static NAN_METHOD(New);
     static Handle<Value> New(mapnik::feature_ptr f_ptr);
+    static NAN_METHOD(fromJSON);
     static NAN_METHOD(id);
     static NAN_METHOD(extent);
     static NAN_METHOD(attributes);
-    static NAN_METHOD(addGeometry);
-    static NAN_METHOD(addAttributes);
-    static NAN_METHOD(numGeometries);
+    static NAN_METHOD(geometry);
     static NAN_METHOD(toString);
     static NAN_METHOD(toJSON);
     static NAN_METHOD(toWKB);