+++ src/mapnik_map.hpp
@@ -1,7 +1,10 @@
 #ifndef __NODE_MAPNIK_MAP_H__
 #define __NODE_MAPNIK_MAP_H__
 
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wunused-parameter"
 #include <nan.h>
+#pragma GCC diagnostic pop
 
 #include "mapnik3x_compatibility.hpp"
 #include MAPNIK_SHARED_INCLUDE