+++ test/expression.test.js
@@ -25,8 +25,7 @@ describe('mapnik.Expression', function() {
 
     it('should support evaluation to js types', function() {
         var expr = new mapnik.Expression("[attr]='value'");
-        var feature = new mapnik.Feature(0);
-        feature.addAttributes({'attr': 'value'});
+        var feature = new mapnik.Feature.fromJSON('{"type":"Feature","properties":{"attr":"value"},"geometry":null}');
         assert.equal(expr.evaluate(feature), true);
         assert.equal(expr.evaluate(feature).toString(), 'true');
     });